OSS完成，菜品管理API差一个新增，

+++ sky-pojo/src/main/java/com/sky/dto/DishPageQueryDTO.java
@@ -7,9 +7,9 @@ import java.io.Serializable;
 @Data
 public class DishPageQueryDTO implements Serializable {
 
-    private int page;
+    private Long page;
 
-    private int pageSize;
+    private Long pageSize;
 
     private String name;
